separate search destination component
